+++ docs/code/Foundation.js.html
@@ -595,7 +595,7 @@ export default class Foundation extends EventSystem {
 <span class="jsdoc-message">
 	Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.6</a>
 	
-		on 2021-02-02T03:00:00-03:00
+		on 2021-02-02T11:54:18-03:00
 	
 	using the <a href="https://github.com/docstrap/docstrap">DocStrap template</a>.
 </span>
