+++ docs/code/index.html
@@ -98,7 +98,6 @@
 <h2>Links</h2>
 <ul>
 <li><a href="https://web2solutions.github.io/agnostic-web-foundation/code/index.html">Code documentation</a></li>
-<li><a href="https://web2solutions.github.io/agnostic-web-foundation/reports/index.html">Unit tests Report</a></li>
 <li><a href="https://web2solutions.github.io/agnostic-web-foundation/reports/unit-testing/index.html">Unit tests Report</a></li>
 </ul></article>
     </section>
@@ -152,7 +151,7 @@
 <span class="jsdoc-message">
 	Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.6</a>
 	
-		on 2021-02-02T03:00:00-03:00
+		on 2021-02-02T11:54:19-03:00
 	
 	using the <a href="https://github.com/docstrap/docstrap">DocStrap template</a>.
 </span>
