--- src/App_old.js
@@ -1,60 +1,0 @@
-/* globals document, window */
-import React from 'react'
-import Crud from './components/crud/Crud'
-
-import './App.css'
-
-class App extends React.Component {
-  constructor (props) {
-    super(props)
-    this.foundation = props.foundation
-    this.state = {
-      users: []
-    }
-    console.error('START APP')
-  }
-
-  render () {
-    console.error('App render ')
-    return (
-      <>
-        <header>
-          <nav className='navbar navbar-expand-md navbar-dark fixed-top bg-dark'>
-            <a className='navbar-brand' href='#'>Fixed navbar</a>
-            <button className='navbar-toggler' type='button' data-toggle='collapse' data-target='#navbarCollapse' aria-controls='navbarCollapse' aria-expanded='false' aria-label='Toggle navigation'>
-              <span className='navbar-toggler-icon' />
-            </button>
-            <div className='collapse navbar-collapse' id='navbarCollapse'>
-              <ul className='navbar-nav mr-auto'>
-                <li className='nav-item active'>
-                  <a className='nav-link' href='#'>Home <span className='sr-only'>(current)</span></a>
-                </li>
-                <li className='nav-item'>
-                  <a className='nav-link' href='#'>Link</a>
-                </li>
-                <li className='nav-item'>
-                  <a className='nav-link disabled' href='#'>Disabled</a>
-                </li>
-              </ul>
-              <form className='form-inline mt-2 mt-md-0'>
-                <input className='form-control mr-sm-2' type='text' placeholder='Search' aria-label='Search' />
-                <button className='btn btn-outline-primary my-2 my-sm-0' type='submit'>Search</button>
-              </form>
-            </div>
-          </nav>
-        </header>
-        <main role='main' className='container appDrawer'>
-          <Crud entity='User' foundation={this.foundation} />
-        </main>
-
-        <footer className='footer'>
-          <div className='container'>
-            <span id='guid' className='text-muted'>Place sticky footer content here.</span>
-          </div>
-        </footer>
-      </>
-    )
-  }
-}
-
-export default App
