--- src/index.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-    <title>Agnostic Web Foundation</title>
-    <meta property="og:site_name" content="Agnostic Web Foundation" />
-
-    <link rel="manifest" href="manifest.json">
-    
-    <meta name="apple-mobile-web-app-title" content="Agnostic Web Foundation" />
-    <meta name="apple-mobile-web-app-status-bar-style" content="white" />
-    <meta name="apple-mobile-web-app-capable" content="yes">
-    <link rel="apple-touch-icon" href="/static/icon_x192.png">
-    
-    <meta name="robots" content="max-snippet:20, max-image-preview:large">
-    
-    <meta name="theme-color" content="#ffffff" />
-    
-    <meta name="description" content="Agnostic Web Foundation">
-    <meta name="author" content="Eduardo Perotta de Almeida">
-    
-    <link rel="preload" href="main.js" as="script">
-    
-    <meta name="mobile-web-app-capable" content="yes">
-    
-    <meta name="application-name" content="Agnostic Web Foundation">
-    <meta name="apple-mobile-web-app-title" content="Agnostic Web Foundation">
-    
-    <meta name="msapplication-starturl" content="/index.html">
-
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
-	<body>
-    <div id="root"></div>
-    <noscript>
-      <style>
-        body {
-          -ms-overflow-style: scrollbar;
-          overflow-y: scroll;
-          overscroll-behavior-y: none;
-        }
-
-        .errorContainer {
-          background-color: #FFF;
-          color: #0F1419;
-          max-width: 600px;
-          margin: 0 auto;
-          padding: 10%;
-          font-family: Helvetica, sans-serif;
-          font-size: 16px;
-        }
-
-        .errorButton {
-          margin: 3em 0;
-        }
-
-        .errorButton a {
-          background: #1DA1F2;
-          border-radius: 2.5em;
-          color: white;
-          padding: 1em 2em;
-          text-decoration: none;
-        }
-
-        .errorButton a:hover,
-        .errorButton a:focus {
-          background: rgb(26, 145, 218);
-        }
-
-        #placeholder,
-        #react-root {
-          display: none !important;
-        }
-      </style>
-      <div class="errorContainer">
-        <h1>JavaScript is not available.</h1>
-        <p>We’ve detected that JavaScript is disabled in this browser. Please enable JavaScript or switch to a supported browser to continue using this application. You can see a list of supported browsers in our Help Center.</p>
-        <p class="errorButton"><a>Help Center</a></p>
-      </div>
-      </noscript>
-      <script type="text/javascript" src="main.js"></script>
-      
-
-	</body>
-</html>
